crop_backend/analysis_worker/requirements.txt

pika==1.3.2
requests==2.31.0
psycopg2-binary==2.9.9
